:truck: Move folders to "pages"
